Set the URL to include the modified query. Part of STRIPES-85.
At present we express URLs in the form:
	http://localhost:3000/users/water
with the query in the path rather than in the URL query, as:
	http://localhost:3000/users?q=water

That's just because at present Stripes Connect can't get at the URL
query, and we want to show searching working from top to bottom. Once
Stripes Connect is up to speed (probably towards the end of next week)
we will change the URLs used for searching.

Because that change is coming, I am not worrying too much in the
short term about this wrinkle: we currently set the whole of the path
to "/users/"+value, so it only works when the Users module is mounted
on /users. But since that's always, I am not losing sleep.