*Temporary* fix: refresh permissions when changing viewed user
It turned out that componentWillReceiveProps wasn't getting called for
<UserPermissions> when we click onto a different user -- presumably
because, although the path changes, the props don't. (This doesn't
make much sense to me, as the 'match' object changes to reflect the
new userid, but maybe such a deep change is somehow missed.)

Just for now, we redundantly pass {...this.props} into
<UserPermissions>, just so that componentWillReceiveProps fires. But
we should figure out properly what is going wrong here.