Add temporary ESLint rule to allow use of the "_id" indentifier
Patron-Group objects carry their ID in a field called "_id" (with a
leading underscore); whereas other kinds of record, such as User and
Item, carry theirs in "id". Unless there is a good reason for this
variation, the partron-group record format should be changed to use
"id" like everything else. Once this is done, we should get rid of the
new "no-underscore-dangle" rule.