Do not pass viewUserProps into userPerms mutators
It turns out that we didn't need this at all: the relevant bits of the
react-router location object are already available down inside
stripes-connect. (I admit I don't really understand why it works -- I
was expecting to have to dummy up a fake location that carried the
relevant path-placeholder parameter, but apparently not.)

So we don't need to pass this.props down from ViewUser.

Fixes STRIPES-280.