Change permissions checks to work with folio-backend-auth v0.9.3
This is related to the permissions changes in DMOD-180. But it seems
that v0.9.3 of the backend VM caught this work in an unfinished state,
so that the permissions it implements are different both from those
that ui-users previously checked and those documented in DMOD-180.

As a result, the changes that have been made in this commit will need
to be changed again when we move to a subsequent folio-backend-auth
VM.

The changes are as follows:

Old permission       New permission            For next backend VM
------------------------------------------------------------------
users.create         users-bl.createuser       users-bl.item.post
users.edit           users-bl.edituser         users-bl.item.put
users.read.basic     users-bl.viewuser         users-bl.item.get
perms.users.read     perms.users.get           [same]
perms.users.delete   perms.users.item.delete   [see (*) below]
perms.users.modify   perms.users.item.put      [same]

	(*) Heikki's DMOD-180 comment says that perms.users.item.put
	should control "Can assign and unassign permissions". That
	seems wrong: adding a permission to a user and removing one
	are rather different operations -- in particular the former
	can escalate privilege, but the latter cannot.

Meanwhile, the changes in lib/RenderPermissions/RenderPermissions.js
may conflict with work to be done for STRIPES-301, so we'll need to be
careful in merging these two change-sets.

For now, this fixes STRIPES-304, but we will need to re-open that
issue when the new backend VM comes along.