New users get perms.permissions.get permission.
This is what replaces the perms.permissions.read permssions that was
causing all the trouble in DMOD-240. We do need this, so that when we
load a user, we have the list of avaliable permissions that could be
added to the user.

(We SHOULD not need it, though. See STRIPES-207. When that is done,
this change can be reverted.)