Ability to add perm to user based on perms.users.item.post
Not on perms.users.item.put, as previously. I determined
experimentally that this is the permission that the backend actually
requires. Which is fair enough, since the actual operation is an HTTP
POST rather then a PUT. (I am not really sure what a PUT could mean in
this situation.)

(The table of new permission names in DMOD-180 said to use
perms.users.item.put; I think that must just be a mistake.)

This constitutes the last part of STRIPES-304 -- the Users app now
correctly handles permissions, and will continue to do so right up
until the next release of folio-backend-auth when it will all break
again.

"Move fast and break things" ... yeah. Not so much fun when you're the
one sweeping up the pieces.