Require three permissions to display "New user" button
We now require
* users.item.post -- to create the new user record
* login.item.post -- to set the new user's credentials
* perms.users.item.post -- to give the new user initial permissions

Unless the user has all three of these, it will not be possible to
complete the create-new-user sequence, so the button must be hidden
unless they are all present.

In a saner world, of course, the Business Logic module would provide
an all-in-one create-new-user API, controlled by a single permission
users-bl.user-create, to which we would pass all the relevant
information in a single shot.