Require perms.users.item.put for adding OR deleting a permission from a set
It turns out that @perms.permissions.item.post" and
"perms.permissions.item.delete" control the ability to create and
delete entire permission sets, not the permissions that they
contain. Changing the set of contained permissions is done by PUTting
a whole new permission-set object, so the same permission controls
both operations.

At present, that permission is perms.users.item.put. This is nothing
to do with users, so it seems to me that the relevant permission is
perms.permissions.item.put. This may change (*sigh*) in the future,
but for now I will use the one that is actually implemented.