Fix a rendering bug that is provoked by ongoing stripes-connect changes
This is subtle, and I am not sure I fully understand
it. resultsFormatter referred in a closure to `data`, which had been
set to `this.props.data`. In certain circumstances -- for example,
when you have viewed the Users app, then switched to Trivial, then
switched back -- `data` would still be undefined, even though
`this.props.data` had since changed and contained data. The result was
that records would be renderd, the formatter would be invoked, and it
would try to look up patron-groups in the old undefined data.

Thw fix is to have the closure refer explicitly to `this.props.data`
rather than just to `data`. But how does the former get a chance to
change before the formatter is invoked? I don't know.