clear-search button goes back to Users starting state
As opposed to the root of the application. This means that now only
enabled users are included in the list, and they are sorted by name.

Fixes UIU-155.

However: the UI state (filters, highlighted sort criterion) are not
changed to match what is specified in the new URL. That is a distinct
problem with potentially significant ramifications, which I shall file
as a separste issue.