better counting of user's loans.
Retrieve loan counts by running a limit=1 query for each of open and
closed loans, which incidentally include metadata about the total number
of items that would be returned absent the limit clause. This is not the
most efficient implementation ever, but it's vastly better than the
previous approach of retrieving ALL details about ALL loans and then
filtering that list to generate the counts.