Move "userCount" resource back into <Users>
This make infinite scroll work, though somewhat inelegantly: the
`userCount` resource is now back in `<Users>`, where it needs to be
for stripes-connect to spot it changing and re-load the `users`
resource; and `<SearchAndSort>` manipulates its value using a new
`parentMutator` property (which I guess we were going to need later
anyway, so I don't feel _too_ bad about it.)

Future refinements are discussed in UIU-286.

Fixes UIU-284.