Tentative work to move the create-new-record code
I'm moving this back out of the generic <SearchAndSort> component into
<Users> because it's so very user-specific: the initial user-record
creation, for example, looks for and massages addresses, and then of
course it goes on to make a "credentials user" and a "permissions
user".

This is on a sub-branch rather than the main UIU-299 branch because I
think the current state of the code has lost some functionality: newly
created users are not shown in the user list, nor highlighted. If I am
right in thinking that the current UIU-299 code does have this
functionality, I need to figure out how I lost it.