Do not pass rowMetadata into <MultiColumnList>
I want to dump this because it contains another hardwired reference to
Users -- specifically, to the "username" field.  But is the
`rowMetaData` property used at all in `<MultiColumnList>`? From
code-reading, it doesn't appear to be. It only ever gets passed as the
third argument to a function that has only two parameters.

And experimentally, simply not passing it in doesn't seem to prevent
anything from working.