SearchAndSort's show-single-row is now optional
The desired behavior of this app when used as a plugin as via
ui-plugin-find-users is different than when used as a standalone app. On
its own, a search that results in a single row should show that entry's
details pane. As part of a plugin, when being displayed in a modal, a
search that results in a single row should simply show that single row.

This change restores the previous functionality of the app when used as
a plugin while still allowing the app to take advantage of
SearchAndSort's show-single-row feature when used independently.

Refs STSMACOM-52, UIREQ-60, UICHKOUT-54, UIU-373, UIIN-58.