lift loansHistory state; towards UIU-419
The way this was coded previously, the loans history query would not be
run if the loan-details page was accessed directly; it was only run
on the loan-list page. This means that when the loan-details page was
loaded from external applications or bookmarks, the page would load but
it would not show any details.