ask for extra facet rows.
Oh, wicked, bad, naughty, evil hack. We are rather abusing facets here
to do some reporting on group member-counts, when really facets were not
designed for this.

By default, facets return only the top five results by count, which is
probably fine if you're actually doing a faceted search and want to find
the most relevant results. But this isn't searching; it's reporting, and
we want to retrieve every single row. Unfortunately, we don't have
another way to get at the group member-counts at present, so we abuse
the desired-rows parameter to the facet and hope we don't ever have more
than 50 patron groups.

This is not a long-term fix.

Refs MODUSERS-57.