pass table-header keys, not values to plugin
In order to handle internationalized strings in a table-header, we need
to always pass in a map of field-names to header-values rather than
relying on the field-name to match the header-value. The field name,
e.g. username, patronGroup, etc. is never displayed whereas the
header-value, e.g. Username, Patron Group, is.

The header-name may be translated and therefore cannot be used as a key.

Fixes UIU-452