disable username validation test
When Electron is automated through NIghtmareJS, sometimes the async
validation on the username field just doesn't run, as though the blur
event never happens. Instrumenting the code shows this is literally what
happens; sometimes the event is simply never triggered.

With --show, I can see it get stuck and reliably cause it to get unstuck
by clicking into alternate fields, but automating these same clicks via
Nightmare does not work.

Why is this failing now, when the changes related to user-status filters
are entirely unlrelated? I have no idea.

This be madness, and there is NO method in't.

Refs UIU-400