avoid validation NPE
Validate was expecting to retrieve an item from a list and then retrieve
a property from that item but not considering the possibility that the
list was empty. This change avoids the null pointer exception in order to
keep the UI from becoming corrupted but does address the underlying
cause. Consequently, validation when the list contains no items always
fails, but at least it doesn't explode when it fails.