update tests to reflect changed attributes
Many attributes in `<EntryManager>` have changed, principally removal of
the title attribute, so the updated tests reflect that. `xclick` remains
a bit dodgy and has been largely removed since if it fails it seems to
leave the Electron process hanging around and leads to weird failures
and exceptions on later tests. Finally, the optional address section of
the new-user form is now omitted in the new-user test since that test
relied on some unreliable logic to find the "add" button.