don't use FormattedMessage for options
The label for an `<option>` must be a string so we need to use
`formatMessage()` instead of `<FormattedMessage>`. This reverts that
portion of the work in #574.

Additionally, it sets the values for the `<option>`s in the
proxy-related fields to hard-coded strings consistent with the immutable
values on the backend.

Fixes UIU-736