correctly display proxy/sponsor relationship date
When a proxy or sponsor is newly-added, the `proxy.metadata.createdDate`
field is empty. Sending `null` to `<FormattedTime>` results in a display
corresponding to the start of the unix epoch, midnight on 1970-01-01
whereas `stripes.formatDateTime` returned an empty string. This change uses
`<FormattedTime>` but correctly checks the value of the field before
displaying it, fixing the bug introduced in #574.

Refs UIU-679