support updated okapi interfaces
New features were added in users-bl that necessitated bumping its
interface, but without support for the new interface in the UI, we
couldn't build folio-snapshot. Providing parallel support for the old
interface (since no functionality here is changing) and the new
interface (so that backend dependencies can be satisfied) seems like the
least painful solution.

Refs FOLIO-1735, MODLOGIN-103, MODUSERBL-62