support "Open - In transit" status
Outbound links to the requests app now include the filter "Open - In
transit" in addition to other "open request" fiters".

Additionally, outbound links to the requests app exactly match the case
of the filters, causing the filter-checkboxes to properly appear in the
checked or unchecked state.

Additionally, queries for open or closed requests within this app to
use a substring match for "open" or "closed" rather than substring
matches on the full filter, e.g. "Open - In transit" which will
automatically handle additional statuses beginning with "Open" or
"Closed" in the future.

Fixes UIU-682