always show explantion for renewal failure (#795)
The bug report here was that when a renewal was denied to the item
having an open recall-request against it, the explanation was missing
from the alert box.

This occurred because the error handler in `withRenew` inadvertently
returned a `<FormattedMessage>` with nested props when the error
contained a policy-name, which it does when the loan-policy is the
reason for the denial. When the error is the result of an open recall,
no policy-name is present, hence no nested props.

This fix cleans up the error handling in `withRenew` so the return value
is consistently structured for an error of any type, and it updates the
explanation display accordingly.

This bug was introduced in #600 during internationalization when
the code changed from something like

    message = `could not renew because ${message}`;
    if (policyName) {
            message = `${message}; see ${policyName}`;
    }

to

    message = <FormattedMessage key="foo" values={{message}} />;
    if (policyName) {
        message = <FormattedMessage key="bar" values={{message, policyName}} />;
    }

Reusing the message variable was fine in first case when it always resolves
to a string. After refactoring, it is passed as a string to the first
`<FormattedMessage>` object but then is passed as a `<FormattedMessage>`
to the second `<FormattedMessage>` object, causing the nested props.

Fixes UIU-971