fix new-permission-set test
This test relied on `xclick`, which was removed from stripes-testing in
`v1.6.0`, which means that should've been tagged `v2.0.0`, but here we
are. `xclick` has been a known bad actor for ages now and should have
been removed from this test long ago.

In support of removing `xclick`, `data-` attributes were added to both
permission-add buttons and the permissions-in-this-set list in order to
provide unique selectors, and the non-unique `id` attribute that was
used for _every_ permission's delete button was made to be unique,
because, well, that's how `id` attributes are supposed to work.