UIU-820 allow proxies with empty expiration dates
When checking whether a proxy relationship should be valid, an account
with no expiration date should be considered valid. The previous version
fed the empty value to `moment`, which I think resulted in moment
deciding that an empty date was "before" the current date, which caused
a validation error.

Fixes UIU-820