STCOM-571 include latest stripes-components
PR #950 included changes to the add/edit user form's footer that relied
on a change to stripes-components that was not present in stripes
v2.10.0. That change was added in a patch release, v2.10.1, so this PR
updates the minimum stripes version to v2.10.1.

Yarn will do The Right Thing with the `~` dependencies here and pull in
that patch release automatically, but if any platform build processes
remove that `~` and use straight version, then this will protect them
and cause the correct version to be pulled in.

Refs STCOM-571