CHAL-30 more efficient request retrieval
Tweak the query that retrieves open requests related to a patron's loans
to be slightly more efficient. This is not a huge improvement, though
it's a little bit that can add up over many loans. The real problem is
that the requests API does not allow for the retrieval of requests keyed
by the borrower's ID, only the item ID, so the only way to find requests
for a patron's loans is to pass in the item ID for every loan.

Refs CHAL-30