pin stripes to v2.8.0
Given that this will be deployed to an environment with stripes v2.8.0
and that it is based of a release (v2.4) when that was the current
stripes release, we will pin the dev-dep on stripes to v2.8 as well.

The issue here is that subtle differences in the markup of some
components, likely `<MultiColumnList>`, is causing unit test failures
because we have v2.8-era selectors in the tests here in ui-users but NPM
is pulling down a newer version to conduct the tests. At the API level,
`<MCL>` _is still compatible_, but its internal markup, which these
selectors inadvertently rely on, has changed.