CHAL-30 retrieve requests-against-patron in batches
The structure of the requests API is such that there is no way to
generate a query that retrieves open requests against a user given only
that user's ID. Instead, we must provide the item-IDs of the user's open
loans and retrieve requests based on those item-IDs. For a user with
many loans, this generates a very long querystring, one that is, in
fact, _too_ long, causing Okapi to return a "414 Request URI Too Long"
response.

The approach here is to implement paging, retrieving potential requests
in batches of 50. It's a bit of a hack, but it's not egregious. I picked
50 because I know a user with 100 loans causes the error and this was
halfway. We could probably bisect further and arrive at a higher number
between 50 and 100 but I'm not sure it's worth the effort.

Refs CHAL-30