appease Nightmare's mysterious desires
There's a bit of reorganization of the form-fields for clarity here, so
the order they appear in the test is more consistent with the order
they appear on-screen, and there's the new, required preferred-contact
field.

But the interesting thing here is that this test failed _not_ because it
was missing the required preferred-contact field but because the click
sent to the dummy-search:
```
-          .type('#input-user-search', '0')
-          .wait('button[type=submit]')
-          .click('button[type=submit]')
```
somehow caused future interaction with the user-form to immediately
close the form and cause nightmare to get lost in neverneverland on the
user-list page.

This makes _no sense whatsoever_.

The user-search form is entirely separate from the new-user form, and
even if a click on the form somehow registered as a click on the latter,
the new-user form should fail validation because no data has been
entered yet.

This is not unlike the weird, weird stuff going on in the new-proxy
integration test in platform-core where clicking a row in the results of
a browse in ui-plugin-find-user does ... something else and causes the
user to logout.

Gimme a W!
Gimme a T!
Gimme an F!
What's that spell?
"Nightmare dot JS!"