UIU-897 props cleanup (#1015)
It's not really fair to pin this on UIU-897 since many of these errors
were present previously. All that's going on here is some cleanup of
incorrect and/or missing and/or extraneous props.

* `EditUserInfo` needed a formatter for its boolean status value
* `UserEdit` omitted `defaultWidth`
* `UserForm` included `stripes` but that's handled by `withStripes`

I found this stuff while pursuing something else that was blowing up in
the console, and the console pollution from this stuff was just killing
me, so I fixed it.