UIU-1380 don't search items by barcode wildcard
Searching for items by barcode wildcard is not performant on a
production system with millions of entries. The intent here is that a
barcode will be scanned in so there is really no need for a wildcard
search anyway; this is essentially "lookup by barcode" not "search by
barcode" and the query now reflects that.

Refs UIU-1380