FOLIO-2097 avoid CommandList and HasCommand
We have a theory that a memory leak in a transitive dependency of
`<CommandList>` and `<HasCommand>` is behind the failing unit tests. The
problem may be that event listeners are not properly detached and
recycled.

This is a test of that. Obviously, we don't actually want to get rid of
key bindings. But we would really REALLY like to have reliable unit
tests.

Refs FOLIO-2097