FOLIO-2097 comment CommandList and HasCommand
We have a theory that a memory leak in a transitive dependency of
`<CommandList>` and `<HasCommand>` is behind the failing unit tests. The
problem may be that event listeners are not properly detached and
recycled.

This is an alternative to #1139, commenting out those components instead
of removing them, which may be easier to revert.

Refs FOLIO-2097