UIU-1406 correctly sort "Active" column
The first column in the search results had the name `status` but the
field-name on the object is `active`. I think we have gone around and
around on how to properly display this field. In any case, the problem
here is that there is no `status` field to sort by, so of course trying
to use that does not work.

Fixes UIU-1406