UIU-1349 Navigate back to users preview/search from edit view on cancel (#1136)
## Problem
UI-users stuck in a navigation loop if another app is visited while the edit/create screen is open. Clicking the X or Cancel buttons would use `history.back()` which could send the user back to an outside app.

## Approach
Navigate to the users' detail view by `history.push()` rather than depending on `history.back()`. The search portion of the location is retained and added to the pushed URL so that search results/filters/text will be kept.

fixes UIU-1349

* navigate back to users preview/search from edit view on cancel
* use search for navigation after successful save