UIU-1457 restore CommandList, and HasCommand
`<CommandList>` and `<HasCommand>`, due to a transitive dependency on
`mousetrap`, leaked memory which caused problems in CI when running unit
tests.

This branch restores those wrappers.

Once the change in `stripes-react-hotkeys` is merged and the version is
updated in `stripes-components`, we can update the version in stripes
and then here in `package.json`.

Refs UIU-1457, STCOM-636