correctly restrict stripes-core devDep
PR #1178 changed the dev-dep on `stripes-core` from caret to tilde
without explanation. I think this was unintentional, but it had the
effect of pulling multiple copies of `stripes-core` into the build as
soon as the next minor version of `stripes-core` was published. This
happened because the looser caret dev-dep on `@folio/stripes` pulled
in the new version while the tilde dev-dep specified here held on to
the old version.