separate section-settings to avoid dep loop
`index.js` and `settings/index.js` were circularly dependent, but only
in order to get at the `settingsSections` array that was exported from
`index.js` _in addition_ to its default export of `UsersRouting`. Sure
it's possible to export multiple things in this manner, but exporting a
class and an array from the same file is icky. It's not like we're
exporting a dozen functions from a utility library here.