UIU-1544 - updated recordId entity in manifest (#1217)
It's very confusing how to use a local resource that should evaluate to a scalar.
A resource is an object. It may have scalar values, but the resource itself is an 
object. 

Here, trying to embed the resource `%{recordId}` in the path caused an error
because `recordId`, i.e. the resource itself, is an object, and thus would be 
serialized to `[object Object]`. The documentation isn't great on how this is
supposed to work. 

Fixes UIU-1544