stripes devDeps must use caret, not tilde (#1226)
`yarn` will load deps based on the _loosest_ dependency it finds, not
the strictest. This means if you have transitive deps on `~1.2.3` and
`^1.2.3` and the available package is `1.3.0` you'll get TWO copies!
Hooray! Poop.

We only want one copy of each of the `@folio/stripes-*` libraries in our
builds. With multiple copies, unit tests will not run.

Reverts that change from #1195.