remove unused checkScope function
This function got added, twice, _months_ ago in PR #839, but SonarCloud
just started complaining about it now. I don't know why it only came up
now, but it is (1) clearly a duplicate function, which is a bug and (2)
not called anywhere, which means it's safe to just get rid of it. So I'm
getting rid of it.