UIU-488 allow for localizable permission names
At present, we pull the display names for permission sets out of the
structure defined in `package.json` which doesn't allow for them to be
localized. Instead, the `permissionName` should be used to construct a
translation key and permission names should be displayed just as are any
other embedded value.

Refs UIU-488