update MCL anchor handling
New-permission set form used to have a selector like `a div` but it
appears that switched to `div a` ... somewhere, probably in
`<MultiColumnList>` in stripes-components. This update tries to make the
selectors a little bit more semantic/robust. I'm not sure if it actually
does that or if it just rearranges things, but in any case tests work
again, so, yay.

Also, this moves all clicks into nightmare instead of doing some in
browser-scope, which is much more reliable.

See also folio-org/stripes-testing/pull/69 for parallel changes in the
shared handlers.