Monkey patch BigTest (#1285)
Monkey patch BigTest's (Interactor.prototype.run) in order to turn off the famous error:
```
convergent assertion was successful, but exceeded the ${timeout}ms timeout.
```
The important change is here:

https://github.com/folio-org/ui-users/pull/1285/files#diff-6dd65ac431218c762d36549899c792ffR92-R108

Additionally, bump up the timeout on a few individual tests that are failing to pass. 
It may be that these timeouts will be unnecessary once the patch is merged, but 
until then, we need more time to run these multi-step tests. 

Co-authored-by: Zak Burke <zburke@cornell.edu>