STRIPES-678 pin moment to ~2.24
Pin `moment` at `~2.24.0` in light of multiple issues with `2.25.0`
([5489](https://github.com/moment/moment/issues/5489), [5472](https://github.com/moment/moment/issues/5472)).

The `resolutions` entry is necessary because `2.25` leaks in via
transitive deps. I don't know why it is necessary here when it isn't,
say, in ui-circulation, which pulls it in via the same route of
stripes-cli and stripes-core. That makes no sense. Alas.

Refs [STRIPES-678](https://issues.folio.org/browse/STRIPES-678)