UIU-496 match protocol in remote images
Images pulled from remote sites should match the protocol used to load
the parent page in order to avoid security warnings in the console.

The ticket mentions that a local image should be used in case
"placehold.it is down" which is legit objection but it's not worth the
effort IMHO as it in no way affects functionality.

Refs UIU-496