UIU-1672 be dumber about the "reset password" link
Previously, the `CreateResetPasswordControl` component would retrieve
the credentials of the given user in order to investigate the hash to
see if it looked like an empty string, because new users were given
empty string as a password, in order to choose whether to show the
"create password" or "reset password" modal.

There are two big problems with that: (1) we were creating users with
empty string passwords which should not be allowed, and (2) we were
able to retrieve a password's hash and salt which should not be allowed.
So assuming we're going to stop creating users in that manner, and that
we're going to stop allowing the retrieval of password hashes and salts,
the easiest thing to do here is to just show the "reset password"
component and not worry about "reset" vs "create".

Refs UIU-1672, UIU-1671, MODLOGIN-128