UIU-1671 do not create empty credentials records (#1357)
Previously, the "reset password" functionality of mod-login would fail
if a credentials record for the given user did not already exist
(MODLOGIN-131). A hedge around this was to create credentials records
with empty-string passwords whenever a username was added to a user
record. But that's a giant security hole, and now that MODLOGIN-131 is
resolved we no longer need to keep that workaround in place.

Refs UIU-1671, MODLOGIN-131