UIU-681: QueryValidationException: CQL: only right truncation
QueryValidationException: CQL: only right truncation supported for id: *foo*

I addessed this by stripping out asterisks in he "makeQuery" function-
this should remove the character as the query is being formatted, so the
back-end won't see it.  I'm also stripping asterisks out in the
querySetterfunction so that the URL and the search field will reflect
a search without an asterisk, so the user can see it was stripped out.

refs: https://issues.folio.org/browse/UIU-681