UIU-1457 restore CommandList, HasCommand (#1389)
`<CommandList>` and `<HasCommand>`, due to a transitive dependency on
`mousetrap`, leaked memory which caused problems in CI when running unit
tests.

This branch restores those wrappers which were removed in #1139.

Refs [UIU-1457](https://issues.folio.org/browse/UIU-1457), [STCOM-636](https://issues.folio.org/browse/STCOM-636)

Replaces #1208, which replaced #1161. There's no material difference
among any of these three PRs, but `master` had drifted so much that it
was easier to start from scratch.