UIU-1750: Filter users by tags does not work
The source of this problem was that the tags filter in the user search
was set to use tag ID when filtering by tag. This does not work because
the 'tags' smart component does not store tag IDs in entity records. It
stores the tag names as a list.  So the only way to filter for any entity,
including users, is to filter by tag name.  Changing this would require
reworking a fairly substantial portion of the tags smart component, so I
simply changed the search in ui-users back to filtering by tag name.

refs: https://issues.folio.org/browse/UIU-1750