UIU-1801: crashing when accessing the "edit user details" screen directly
With some help, worked out that this was caused by the form rendering
before the data fetch operation was complete, the form would then
crash because it had no user data to display.  Added a check in render
that looks to see if the edit is on an existing user, and wether the
data fetch is done.  Also rewrote all test changes so that tests
can now visit the user edit screen directly.

refs: https://issues.folio.org/browse/UIU-1801