consistent spacing around the colon
The output we are after here is something like this:
```
Barcode: <a href="">1234</a>
```
but the way it was coded, the translation value needed to contain
`Barcode: ` (note the trailing space). Sometimes it did, sometimes it
didn't, sometimes it didn't have the colon either. It was amazing how
much inconsistency there was, and how much that one space bugged the
crap outta me. No longer. No whitespace and no punctuation in the
translation should bring more consistent results.