UIU-1226 Claimed returned in-app report: Existing fields
The existing loans report had a lot of code that I would have duplicated,
so rather than write new code that runs a separate claimed returned report,
I altered the loan report code to make it more generic.  When you call the
"generate" method, it now takes a string that tells it what kind of report
you want to run.  The report generator sets the columns and query string
appropriately, then retrieves the data.  I moved the column names into a
separate file and separated them out-right now, the columns for the two
reports are identical, but I can see from the story that later we'll want
to add specific columns to the claim returned report-that should now be easy.
I shuffled around a lot of labels that both reports use so their names aren't
as tied to the overdue report.

refs: https://issues.folio.org/browse/UIU-1226