UIU-1045 show "awol loan" when loan is ... awol
If a URL for a loan is bookmarked (or copied, or stored in any way,
really), and then that loan is anonymized, show a "can't find that loan"
message.

Previously, the loan-details page would instead show a "loading"
spinner, as though the loan were being retrieved and just hadn't loaded
yet. Now, we show a "missing loan" message when fetching a user's loans
has finished and no matching loan is present.

Refs UIU-1045