UIU-1214 Claim returned-1: effect on SET COST fees for item status Declared Lost (#1379)
Claim Returned: effect on SET COST fees for item status Declared Lost

Patrons may think they have returned items that are still showing as
checked out on their account. When this happens, the item is still
considered loaned to the patron, but the library will treat the item
differently from other items on the patron's account (the item will not
be considered overdue, and it will be searched for extensively). A
patron should not be charged for an item that they claim they've
returned until the library takes further action (described in separate
stories).

Refs UIU-1214