UIU-2132 include permissions on some search container resources (#1710)
Some search-container resources where were lookup tables for detail
records did not contain permissions, causing errors if the user lacked
the requisite permission.

One approach would be to set `throwErrors` to `false` for these routes;
another is to add permissions; still another would be to separate the
resources fetches into components that are only rendered when the
required permissions are in place. These last two are more or less
equivalent, and while separate components would be still more
self-contained, this is the smallest change possible that still gets to
the correct behavior.

Refs UIU-2132